pyproject.toml
[project]
name = "werewolf_discord"
version = "0.1.0"
description = "A werewolf game for Discord"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5",
    "python-dotenv",
]

[tool.basedpyright]
typeCheckingMode = "recommended"
